split up resolvers and queries
