+++ package.json
@@ -18,8 +18,10 @@
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "graphql": "^0.13.2",
     "graphql-import": "^0.5.0",
     "graphql-request": "^1.5.1",
+    "graphql-tools": "^3.0.0",
     "graphql-yoga": "^1.9.2",
     "pg": "^7.4.1",
     "reflect-metadata": "^0.1.12",
