+++ src/modules/register/register.test.ts
@@ -1,7 +1,6 @@
 import { request } from "graphql-request";
-
-import { User } from "../entity/User";
-import { startServer } from "../startServer";
+import { startServer } from "../../startServer";
+import { User } from "../../entity/User";
 
 let getHost = () => "";
