+++ src/modules/register/schema.graphql
@@ -1,5 +1,5 @@
 type Query {
-  hello(name: String): String!
+  bye: String
 }
 
 type Mutation {
