+++ src/startServer.ts
@@ -1,14 +1,24 @@
 import { importSchema } from "graphql-import";
 import { GraphQLServer } from "graphql-yoga";
 import * as path from "path";
+import * as fs from "fs";
+import { mergeSchemas, makeExecutableSchema } from "graphql-tools";
+import { GraphQLSchema } from "graphql";
 
-import { resolvers } from "./resolvers";
 import { createTypeormConn } from "./utils/createTypeormConn";
 
 export const startServer = async () => {
-  const typeDefs = importSchema(path.join(__dirname, "./schema.graphql"));
+  const schemas: GraphQLSchema[] = [];
+  const folders = fs.readdirSync(path.join(__dirname, "./modules"));
+  folders.forEach(folder => {
+    const { resolvers } = require(`./modules/${folder}/resolvers`);
+    const typeDefs = importSchema(
+      path.join(__dirname, `./modules/${folder}/schema.graphql`)
+    );
+    schemas.push(makeExecutableSchema({ resolvers, typeDefs }));
+  });
 
-  const server = new GraphQLServer({ typeDefs, resolvers });
+  const server = new GraphQLServer({ schema: mergeSchemas({ schemas }) });
   await createTypeormConn();
   const app = await server.start({
     port: process.env.NODE_ENV === "test" ? 0 : 4000
