--- src/tests/constants.ts
@@ -1,1 +1,0 @@
-export const host = "http://localhost:4000";
